+++ sdk/keyvault/azure-security-keyvault-secrets/CHANGELOG.md
@@ -2,4 +2,4 @@
 
 ## 1.0.0-beta.1 (Unreleased)
 
-* initial preview
+- initial preview